Extract terminal and browser panes
