+++ docs/DECISIONS.md
@@ -132,5 +132,6 @@
 - Transcript message controls live in `QuillCodeTranscriptMessageView.swift`. User and assistant message bubbles, retry/use-as-draft actions, feedback actions, and the shared transcript copy button should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed. Tool-card and artifact-preview rendering remains the next transcript family to extract.
 - Tool-card rendering lives in `QuillCodeToolCardView.swift`. Tool cards, execution chips and rails, artifact chips/previews, and raw detail blocks should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed and how copy actions are wired.
 - Settings and runtime issue rendering live in `QuillCodeSettingsView.swift`. Authentication settings, developer override controls, Computer Use permission onboarding, diagnostics, and `QuillCodeSettingsDraft` should evolve there; `WorkspaceSwiftUIView.swift` should only present the sheet and apply the resulting settings update.
+- Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
 - The first memories slice is auditable and user-authored. QuillCode loads bounded `.md`, `.txt`, and `.json` notes from `~/.quillcode/memories` and project `.quillcode/memories`, rejects symlinks and paths outside the memory root, truncates oversized notes with visible labels, snapshots loaded memories onto threads, injects them into TrustedRouter as non-command background context, and surfaces the count/sources through the top bar, Memories pane, static HTML renderer, Playwright harness, slash command, and macOS menu bar. Users can explicitly save global memories with `/remember text`, the Add memory command-palette action, or a model-authored `host.memory.remember` tool call when they ask QuillCode to remember a stable preference/fact; saved notes create transcript/tool audit records, refresh memory context immediately, and reject obvious credentials, tokens, passwords, and private keys. Global memories can be forgotten from the Memories pane and produce a transcript notice. Project memories stay read-only in-app because they are project-owned files. Richer redaction review, idle Chronicle jobs, editing UI, conflict resolution, and fully autonomous inference of memories without explicit user intent are separate safety/UX milestones.
